Copy of production that uses staging image
I wish I did not have to do this. It's kind of a hack.

I _wanted_ to use an environment variable to contain the tag, because that's supported by the docker-compose.yml file format (see: https://docs.docker.com/compose/environment-variables/#substitute-environment-variables-in-compose-files).

However, the Azure WebApp that we're using for hosting appears to have trouble with that approach, because the tag variable name was in the output.