Fixed build for Ubunutu 20LTS environment
Updated the LibGit2Sharp dependency from 0.26.2 to 0.27.0-preview-0096. This is necessary for Ubuntu 20.04LTS as the 0.26.0 version utilizes OpenSSL 1.0 which is no longer supported. While the build was okay, a vast majority of the tests that utilized the LibGit2Sharp dependency would fail. By updating to 0.27.0-preview-0096, which utilizes OpenSSL 1.1, this is no longer an issue and allows all exiting tests to pass.

Resolves: Issue #206