Dependency dates compare against target date EoD
The monthly analysis dates that are generated for comparisons against the dependency release dates have been changed from the beginning of the day (00:00:00.000) to end of day (EoD) (23:59:59.999.9999). This is done in effort to include the asOf dates in these comparisons. This will also prevent dependencies, released the day of Freshli analysis, from being exclude from analysis.

Due to the nature of this change, Acceptance tests were updated to enable EoD asOf dependency analysis which affected results where a new version was released the day of the target date. (Ex. Analyzing a dependency for 12/01/2019 when version 2.0.1 was released on 12/01/2019).

Unit tests and a new acceptance test was created for this change.

Resolves: Issue #162