Migrates to .NET 5.0
This should address the goals of  #193 (Change build target to .NET 5.0).

This also sets the `rollForward` setting in `global.json` to `latestMinor`. That way, it should make it easier for people who are trying to build the project using slightly newer, but still very likely compatible, versions of the .NET SDK.