Enables Nullability for PythonVersionInfo
Annotates the PythonVersionInfo class to enable nullable references. This clears the warning and indicates intent to work with nullable references. C# 8.0 added this annotation to allow for fine-grain control on the nullability of references. This annotation also allows for the change to the SafeToLower and SafeExtract methods, that can potentially return null, to return Nullable strings which further enables null-safe checking.

See Microsoft's documentation https://docs.microsoft.com/en-us/dotnet/csharp/nullable-references#nullable-annotation-context

Resolves #199