Adds Generic Manifest Implementation
Renamed GenericManifest.cs to ManifestModel.cs and relocated to Models directory.
Renamed GenericDependency.cs to DependencyModel.cs and relocated to models directory.
On of the key components Freshli requires to run for any format is a manifest. The repurposed GenericManifest.cs provides an implementation of AbstractManifest.cs that parses json from the ManifestModel.cs format. Tests have been adjsuted and created to verify this conversion. Currently, AbstractManifest requires the version of a requested package to be singular and without operator as a string. A conversion from the Model's specs from and to string has been created to facilitate this. It might be good to flesh out the AbstractManifest version in the future handle better complexities and cut down on work-arounds like this.