Adjusts date ranges so they start with the date of oldest file change
There was a bug where that was causing the start of the date range to be older than the oldest revision in the file history. Meaning that the file would not have existed at the start of the date range that was being used for analysis. This changes the logic so that the oldest file date is always the starting point of the range, and then range is built in one month increments starting with the start of the month that immediately follows the creation of the first file in the history.