Uses time zone from `asOf` value to compute date range increments
The logic was starting with the oldest file and then incrementing one month at a time from there. The timezone of that value might not match the timezone that has been passed in via `asOf`. With this change, we assume that the caller wants the results to be in the same timezone offset that's specified in the `asOf` value. Before this change, all results would have been in the timezone of that first commit.