Uses a personal access token to ensure that CI runs on changelog updates
The issue that I am running into is caused by the CI checks not running on the
commit that updates the change log. By using a GitHub personal access token to
interact with the GitHub API (which is what Actions uses), then I should be able
to work-around that limitation.