Runs CI on macOS, Linux, and Windows
Description:
Uses the "build matrix" functionality in GitHub Actions to run on the latest
available version of macOS, Linux, and Windows.

Rationale:
We often have team members who work on this project from each of these
environments. We've run into challenges in the past where the tests have not worked in
one of the environments, but they have succeeded in the others. Having
this matrix set up will help us stay in front of future instances of
that problem.

Past issues that we've seen have been related to path separators and
platform specific bugs in third party dependencies such as LibGit2Sharp.

Some items to note about how this is set up:

* Code coverage is collected and in each environment, combined, and then
  the combined coverage data is then sent to Code Climate
* The steps for the operations that appear after the the matrix has run
include some duplication of the steps that run within the build matrix.
One way that this duplication could be removed would be to have one of
the matrix builds upload the built files as artifacts that are reused
later on.