Switches from ApprovalTests to VerifyTests
This replaces, ApprovalTests, the library that's being used for golden
master testing with VerifyTests.

This switch is primarily motivated by challenges
that have popped up because of how the ApprovalTests library formats
the `.received.txt` that it generates. When the file that's generated by
ApprovalTests contains multiple lines, for example when the contents of
an array are being verified, then ApprovalTests uses the newline
character `\n` to separate those lines, even when the test is running on
a Windows OS while the typical line separator is `\r\n`.

A secondary benefit is that the output format used by VerifyTests is
"based on" JSON. Which makes the files much easier to read the very
large amount of data that is being verified by these tests. Hopefully,
this will provide benefits when tracking down bugs that might result in
a single value in the file being incorrect.