Breaks up FileHistoryFinder class
Introduces `FileHistoryFinderRegistry` that's responsible for keeping
track of the diffrent `IFileHistoryFinder` instances that have been
registered.

Renames `FileHistoryFinder` class to `FileHistoryService` to better
match the role that it plays. Also moves time-intensive tasks out of the
class's construct. This is to support the shift to `asnyc` calls that
will come later. Constructors cannot run in an `async` fashion, so
anything that might to be `await`ed on needs to be moved out of the
constructor.