Clarifies usage of FileHistoryFinderRegistry
Before the registy class was introduced, there was an assumption that
the `LocalFileHistoryFinder` was always registered. This default
behavior has been removed. Also took an incremental step towards more
dependency injection. This is causing more setup logic to appear in the
tests. That logic can be centralized at a later time.