Makes progress towards introducing new concepts
When investigating issues with the PHP multicast tests, I realized that
it is very common for manifest information to be spread out across
multiple files, but the implementation had the assumption baked in that
all of the information that we needed could be retrieved from a single
file. This is especially not true with the PHP acceptance tests that we
have, because the package repository information is stored in
`composer.json`, but the dependencies are defined in `composer.lock`.
Both files are needed to correctly grab the right information. And doing
that correctly requires some rethinking.

So an `IManifest` is being added along with an `IManifestHistory`. Also,
`IFileHistorySource` is being added. `IManifestService` and
`IFileHistoryService` have been removed, because
they are not needed anymore
after these changes are baked in.

Everything compiles right now, but there are some interfaces and methods
that do not have implementations. This will be corrected in future
commits.