use version 16.3.0 of verify.xunit in tests
this is a sort of goldilocks zone for this dependency: newer versions change the api, and older versions have an unavoidable missingmethodexception error. unfortunately, some change to the library causes 2.0.0 and 2.0 to swap places in verifier output, resulting in a necessary change to a .verified.txt file. i think this is a reasonable tradeoff.